Add a dummy user with uid 1000
Docker 1.2 cannot map the root user of a container to a non-root user of
the Docker host. It is a planned functionality. So we create a dummy user
with a uid matching the host user. Docker can then be run with this dummy
user and created files will be owned by the host user.